add questionnaire update validation

+++ function-web/src/main/java/com/future/function/web/controller/communication/questionnaire/QuestionnaireController.java
@@ -4,6 +4,7 @@ package com.future.function.web.controller.communication.questionnaire;
 import com.future.function.common.enumeration.core.Role;
 import com.future.function.common.properties.core.FileProperties;
 import com.future.function.model.entity.feature.communication.questionnaire.Questionnaire;
+import com.future.function.model.entity.feature.communication.questionnaire.QuestionnaireParticipant;
 import com.future.function.service.api.feature.communication.questionnaire.QuestionnaireService;
 import com.future.function.service.api.feature.core.UserService;
 import com.future.function.session.annotation.WithAnyRole;
@@ -30,6 +31,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Date;
 import java.util.List;
 
 @RestController
@@ -98,6 +100,16 @@ public class QuestionnaireController {
       QuestionnaireRequest questionnaireRequest, Session session
   ) {
 
+  if(!questionnaireService.validateQuestionnaire(
+      null,
+      questionnaireRequest.getStartDate(),
+      questionnaireRequest.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionnaireDetailResponse.builder().build()
+        );
+    }
     Questionnaire newQuestionnaire = Questionnaire.builder()
       .title(questionnaireRequest.getTitle())
       .description(questionnaireRequest.getDesc())
@@ -133,6 +145,17 @@ public class QuestionnaireController {
     @RequestBody
       QuestionnaireRequest questionnaireRequest
   ) {
+
+    if(!!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaireRequest.getStartDate(),
+      questionnaireRequest.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionnaireDetailResponse.builder().build()
+      );
+    }
 
     return QuestionnaireResponseMapper.toDataResponseQuestionnaireDetailResponse(
       questionnaireService.updateQuestionnaire(
@@ -177,6 +200,17 @@ public class QuestionnaireController {
     @RequestBody
       QuestionQuestionnaireRequest questionQuestionnaireRequest
   ) {
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionQuestionnaireResponse.builder().build()
+      );
+    }
 
     return QuestionQuestionnaireResponseMapper.toDataResponseQuestionQuestionnaireResponse(
       questionnaireService.createQuestionQuestionnaire(
@@ -198,6 +232,17 @@ public class QuestionnaireController {
     @RequestBody
       QuestionQuestionnaireRequest questionQuestionnaireRequest
   ) {
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionQuestionnaireResponse.builder().build()
+      );
+    }
 
     return QuestionQuestionnaireResponseMapper.toDataResponseQuestionQuestionnaireResponse(
       questionnaireService.updateQuestionQuestionnaire(
@@ -252,6 +297,18 @@ public class QuestionnaireController {
       QuestionnaireParticipantRequest questionnaireParticipant
   ) {
 
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionnaireParticipantResponse.builder().build()
+      );
+    }
+
     return QuestionnaireParticipantResponseMapper.toDataResponseQuestionnaireParticipantResponse(
       questionnaireService.addQuestionnaireAppraiserToQuestionnaire(
         questionnaireId, questionnaireParticipant.getIdParticipant()),
@@ -269,6 +326,16 @@ public class QuestionnaireController {
       String questionnaireParticipantId
 
   ) {
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toBaseResponse(
+        HttpStatus.NOT_MODIFIED
+      );
+    }
 
     questionnaireService.deleteQuestionnaireAppraiserFromQuestionnaire(
       questionnaireParticipantId);
@@ -308,7 +375,17 @@ public class QuestionnaireController {
     @RequestBody
       QuestionnaireParticipantRequest questionnaireParticipant
   ) {
-
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toDataResponse(
+        HttpStatus.NOT_MODIFIED,
+        QuestionnaireParticipantResponse.builder().build()
+      );
+    }
     return QuestionnaireParticipantResponseMapper.toDataResponseQuestionnaireParticipantResponse(
       questionnaireService.addQuestionnaireAppraiseeToQuestionnaire(
         questionnaireId, questionnaireParticipant.getIdParticipant()),
@@ -325,10 +402,18 @@ public class QuestionnaireController {
     @PathVariable
       String questionnaireParticipantId
   ) {
-
+    Questionnaire questionnaire = questionnaireService.getQuestionnaire(questionnaireId);
+    if(!questionnaireService.validateQuestionnaire(
+      questionnaireId,
+      questionnaire.getStartDate(),
+      questionnaire.getDueDate())
+    ){
+      return ResponseHelper.toBaseResponse(
+        HttpStatus.NOT_MODIFIED
+      );
+    }
     questionnaireService.deleteQuestionnaireAppraiseeFromQuestionnaire(
       questionnaireParticipantId);
     return ResponseHelper.toBaseResponse(HttpStatus.OK);
   }
-
 }
